requirements.txt
fastapi
uvicorn
openai
qdrant-client>=1.9.0
python-dotenv
pydantic
google-generativeai>=0.5.0
passlib[bcrypt]